requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
easyocr==1.7.0
ultralytics==8.0.196
PyMuPDF==1.23.8
Pillow==10.0.1
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
